Split p95 cost-per-trace tests into their own file
test_reporting.py had grown past CodeScene's function-count threshold
for a single module. TestP95CostPerTrace follows the existing
test_reporting_filters.py precedent of splitting reporting tests by
concern into a dedicated file.
